Unique streaming per request
Creating an actor that streams objects from postgres per request.
This way we have more control over each stream using a stateful actor.

For each stream, we receive or create an UUID, which will be helpful to track actions during the streaming.
The actor's name also receives this UUID and logging inside the actor, also make use of it.

When the stream is done, the actor kills itself, preventing memory leak and so.